+++ rmp/src/encode/mod.rs
@@ -12,8 +12,8 @@ mod ext;
 pub use self::sint::{write_nfix, write_i8, write_i16, write_i32, write_i64, write_sint};
 pub use self::uint::{write_pfix, write_u8, write_u16, write_u32, write_u64, write_uint};
 pub use self::dec::{write_f32, write_f64};
-pub use self::str::write_str_len;
-pub use self::bin::write_bin_len;
+pub use self::str::{write_str_len, write_str};
+pub use self::bin::{write_bin_len, write_bin};
 
 use std::error;
 use std::fmt::{self, Display, Formatter};